requirements.txt
anyio==3.7.1
bardapi==0.1.30
beautifulsoup4==4.12.2
browser-cookie3==0.19.1
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
colorama==0.4.6
deep-translator==1.11.4
google-ai-generativelanguage==0.2.0
google-api-core==2.11.1
google-auth==2.22.0
google-cloud-core==2.3.3
google-cloud-translate==3.11.3
google-generativeai==0.1.0
googleapis-common-protos==1.60.0
grpcio==1.56.2
grpcio-status==1.56.2
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.17.3
httpx==0.24.1
hyperframe==6.0.1
idna==3.4
jeepney==0.8.0
lz4==4.3.2
proto-plus==1.22.3
protobuf==4.23.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycryptodomex==3.18.0
python-dotenv==1.0.0
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
urllib3==1.26.16
